feat: public export of type `PackageJson`
